requirements.txt
chardet==3.0.4
pdfminer.six==20200124
pycryptodome==3.9.7
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sortedcontainers==2.1.0
